feat: add Circuit
