--- .fleet/receipt.json
@@ -1,14 +1,0 @@
-// Project generated by Kotlin Multiplatform Wizard
-{
-    "spec": {
-        "template_id": "kmt",
-        "targets": {
-            "web": {
-                "ui": [
-                    "compose"
-                ]
-            }
-        }
-    },
-    "timestamp": "2025-02-24T04:47:09.921477742Z"
-}
