--- .fleet/settings.json
@@ -1,3 +1,0 @@
-{
-    "editor.formatOnSave": true
-}
