+++ .gitignore
@@ -16,6 +16,3 @@ captures
 !*.xcodeproj/project.xcworkspace/
 !*.xcworkspace/contents.xcworkspacedata
 **/xcshareddata/WorkspaceSettings.xcsettings
-
-// GitHub Actions automatically generate
-composeApp/src/commonMain/composeResources/files/aboutlibraries.json
