+++ README.md
@@ -8,3 +8,17 @@
 https://yuyuyuyuyu.dev
 
 <img width="300" alt="Android main screen screenshot" src="https://github.com/user-attachments/assets/37e0204e-5be1-4e39-839b-53c4e1e49804" />
+
+## Run
+
+Desktop
+
+```shell
+./gradlew :composeApp:run
+```
+
+Web
+
+```shell
+./gradlew :composeApp:wasmJsBrowserDevelopmentRun -t
+```
