+++ composeApp/build.gradle.kts
@@ -1,3 +1,4 @@
+import org.jetbrains.compose.desktop.application.dsl.TargetFormat
 import org.jetbrains.kotlin.gradle.ExperimentalWasmDsl
 import org.jetbrains.kotlin.gradle.targets.js.webpack.KotlinWebpackConfig
 
@@ -5,11 +6,12 @@ plugins {
     alias(libs.plugins.kotlinMultiplatform)
     alias(libs.plugins.composeMultiplatform)
     alias(libs.plugins.composeCompiler)
-
-    alias(libs.plugins.mikepenz.aboutlibraries)
+    alias(libs.plugins.composeHotReload)
 }
 
 kotlin {
+    jvm("desktop")
+
     @OptIn(ExperimentalWasmDsl::class)
     wasmJs {
         moduleName = "composeApp"
@@ -31,6 +33,7 @@ kotlin {
     }
 
     sourceSets {
+        val desktopMain by getting
 
         commonMain.dependencies {
             implementation(compose.runtime)
@@ -40,18 +43,31 @@ kotlin {
             implementation(compose.components.resources)
             implementation(compose.components.uiToolingPreview)
             implementation(libs.androidx.lifecycle.viewmodel)
-            implementation(libs.androidx.lifecycle.runtime.compose)
+            implementation(libs.androidx.lifecycle.runtimeCompose)
 
             implementation(compose.materialIconsExtended)
-            implementation(libs.yuyuyuyuyu.simpletopappbar)
-            implementation(libs.yuyuyuyuyu.createtypography)
-            implementation(libs.mikepenz.aboutlibraries.core)
-//            implementation(libs.mikepenz.aboutlibraries.compose)
-            implementation(libs.androidx.compose.navigation)
+
+            implementation(libs.circuit)
+        }
+        commonTest.dependencies {
+            implementation(libs.kotlin.test)
+        }
+        desktopMain.dependencies {
+            implementation(compose.desktop.currentOs)
+            implementation(libs.kotlinx.coroutinesSwing)
         }
     }
 }
 
-aboutLibraries {
-    registerAndroidTasks = false
+
+compose.desktop {
+    application {
+        mainClass = "dev.yuyuyuyuyu.portfolio.MainKt"
+
+        nativeDistributions {
+            targetFormats(TargetFormat.Dmg, TargetFormat.Msi, TargetFormat.Deb)
+            packageName = "dev.yuyuyuyuyu.portfolio"
+            packageVersion = "1.0.0"
+        }
+    }
 }
