+++ composeApp/src/commonMain/kotlin/dev/yuyuyuyuyu/portfolio/data/models/Product.kt
@@ -1,0 +1,8 @@
+package dev.yuyuyuyuyu.portfolio.data.models
+
+data class Product(
+    val name: String,
+    val description: String,
+    val techStack: Set<TechStack>,
+    val repositoryUrl: String,
+)
