+++ composeApp/src/commonMain/kotlin/dev/yuyuyuyuyu/portfolio/ui/appllist/AppListScreen.kt
@@ -1,0 +1,10 @@
+package dev.yuyuyuyuyu.portfolio.ui.appllist
+
+import com.slack.circuit.runtime.CircuitUiState
+import com.slack.circuit.runtime.screen.Screen
+
+data object AppListScreen : Screen {
+    data class State(
+        val apps: List<String>
+    ) : CircuitUiState
+}
