+++ composeApp/src/commonMain/kotlin/dev/yuyuyuyuyu/portfolio/ui/theme/Color.kt
@@ -1,0 +1,219 @@
+package dev.yuyuyuyuyu.portfolio.ui.theme
+
+import androidx.compose.ui.graphics.Color
+
+internal val primaryLight = Color(0xFF3F5F90)
+internal val onPrimaryLight = Color(0xFFFFFFFF)
+internal val primaryContainerLight = Color(0xFFD6E3FF)
+internal val onPrimaryContainerLight = Color(0xFF254777)
+internal val secondaryLight = Color(0xFF2E6A44)
+internal val onSecondaryLight = Color(0xFFFFFFFF)
+internal val secondaryContainerLight = Color(0xFFB2F1C1)
+internal val onSecondaryContainerLight = Color(0xFF12512E)
+internal val tertiaryLight = Color(0xFF87521B)
+internal val onTertiaryLight = Color(0xFFFFFFFF)
+internal val tertiaryContainerLight = Color(0xFFFFDCC0)
+internal val onTertiaryContainerLight = Color(0xFF6B3B03)
+internal val errorLight = Color(0xFFBA1A1A)
+internal val onErrorLight = Color(0xFFFFFFFF)
+internal val errorContainerLight = Color(0xFFFFDAD6)
+internal val onErrorContainerLight = Color(0xFF93000A)
+internal val backgroundLight = Color(0xFFF9F9FF)
+internal val onBackgroundLight = Color(0xFF191C20)
+internal val surfaceLight = Color(0xFFFFF8F8)
+internal val onSurfaceLight = Color(0xFF22191C)
+internal val surfaceVariantLight = Color(0xFFE0E2EC)
+internal val onSurfaceVariantLight = Color(0xFF43474E)
+internal val outlineLight = Color(0xFF74777F)
+internal val outlineVariantLight = Color(0xFFC4C6CF)
+internal val scrimLight = Color(0xFF000000)
+internal val inverseSurfaceLight = Color(0xFF372E30)
+internal val inverseOnSurfaceLight = Color(0xFFFDEDEF)
+internal val inversePrimaryLight = Color(0xFFA8C8FF)
+internal val surfaceDimLight = Color(0xFFE6D6D9)
+internal val surfaceBrightLight = Color(0xFFFFF8F8)
+internal val surfaceContainerLowestLight = Color(0xFFFFFFFF)
+internal val surfaceContainerLowLight = Color(0xFFFFF0F2)
+internal val surfaceContainerLight = Color(0xFFFAEAED)
+internal val surfaceContainerHighLight = Color(0xFFF5E4E7)
+internal val surfaceContainerHighestLight = Color(0xFFEFDFE1)
+
+internal val primaryLightMediumContrast = Color(0xFF103665)
+internal val onPrimaryLightMediumContrast = Color(0xFFFFFFFF)
+internal val primaryContainerLightMediumContrast = Color(0xFF4E6EA0)
+internal val onPrimaryContainerLightMediumContrast = Color(0xFFFFFFFF)
+internal val secondaryLightMediumContrast = Color(0xFF003F1F)
+internal val onSecondaryLightMediumContrast = Color(0xFFFFFFFF)
+internal val secondaryContainerLightMediumContrast = Color(0xFF3E7951)
+internal val onSecondaryContainerLightMediumContrast = Color(0xFFFFFFFF)
+internal val tertiaryLightMediumContrast = Color(0xFF542C00)
+internal val onTertiaryLightMediumContrast = Color(0xFFFFFFFF)
+internal val tertiaryContainerLightMediumContrast = Color(0xFF986028)
+internal val onTertiaryContainerLightMediumContrast = Color(0xFFFFFFFF)
+internal val errorLightMediumContrast = Color(0xFF740006)
+internal val onErrorLightMediumContrast = Color(0xFFFFFFFF)
+internal val errorContainerLightMediumContrast = Color(0xFFCF2C27)
+internal val onErrorContainerLightMediumContrast = Color(0xFFFFFFFF)
+internal val backgroundLightMediumContrast = Color(0xFFF9F9FF)
+internal val onBackgroundLightMediumContrast = Color(0xFF191C20)
+internal val surfaceLightMediumContrast = Color(0xFFFFF8F8)
+internal val onSurfaceLightMediumContrast = Color(0xFF170F11)
+internal val surfaceVariantLightMediumContrast = Color(0xFFE0E2EC)
+internal val onSurfaceVariantLightMediumContrast = Color(0xFF33363E)
+internal val outlineLightMediumContrast = Color(0xFF4F525A)
+internal val outlineVariantLightMediumContrast = Color(0xFF6A6D75)
+internal val scrimLightMediumContrast = Color(0xFF000000)
+internal val inverseSurfaceLightMediumContrast = Color(0xFF372E30)
+internal val inverseOnSurfaceLightMediumContrast = Color(0xFFFDEDEF)
+internal val inversePrimaryLightMediumContrast = Color(0xFFA8C8FF)
+internal val surfaceDimLightMediumContrast = Color(0xFFD2C3C5)
+internal val surfaceBrightLightMediumContrast = Color(0xFFFFF8F8)
+internal val surfaceContainerLowestLightMediumContrast = Color(0xFFFFFFFF)
+internal val surfaceContainerLowLightMediumContrast = Color(0xFFFFF0F2)
+internal val surfaceContainerLightMediumContrast = Color(0xFFF5E4E7)
+internal val surfaceContainerHighLightMediumContrast = Color(0xFFE9D9DC)
+internal val surfaceContainerHighestLightMediumContrast = Color(0xFFDDCED1)
+
+internal val primaryLightHighContrast = Color(0xFF002C5A)
+internal val onPrimaryLightHighContrast = Color(0xFFFFFFFF)
+internal val primaryContainerLightHighContrast = Color(0xFF284979)
+internal val onPrimaryContainerLightHighContrast = Color(0xFFFFFFFF)
+internal val secondaryLightHighContrast = Color(0xFF003419)
+internal val onSecondaryLightHighContrast = Color(0xFFFFFFFF)
+internal val secondaryContainerLightHighContrast = Color(0xFF165430)
+internal val onSecondaryContainerLightHighContrast = Color(0xFFFFFFFF)
+internal val tertiaryLightHighContrast = Color(0xFF452400)
+internal val onTertiaryLightHighContrast = Color(0xFFFFFFFF)
+internal val tertiaryContainerLightHighContrast = Color(0xFF6E3D05)
+internal val onTertiaryContainerLightHighContrast = Color(0xFFFFFFFF)
+internal val errorLightHighContrast = Color(0xFF600004)
+internal val onErrorLightHighContrast = Color(0xFFFFFFFF)
+internal val errorContainerLightHighContrast = Color(0xFF98000A)
+internal val onErrorContainerLightHighContrast = Color(0xFFFFFFFF)
+internal val backgroundLightHighContrast = Color(0xFFF9F9FF)
+internal val onBackgroundLightHighContrast = Color(0xFF191C20)
+internal val surfaceLightHighContrast = Color(0xFFFFF8F8)
+internal val onSurfaceLightHighContrast = Color(0xFF000000)
+internal val surfaceVariantLightHighContrast = Color(0xFFE0E2EC)
+internal val onSurfaceVariantLightHighContrast = Color(0xFF000000)
+internal val outlineLightHighContrast = Color(0xFF292C33)
+internal val outlineVariantLightHighContrast = Color(0xFF464951)
+internal val scrimLightHighContrast = Color(0xFF000000)
+internal val inverseSurfaceLightHighContrast = Color(0xFF372E30)
+internal val inverseOnSurfaceLightHighContrast = Color(0xFFFFFFFF)
+internal val inversePrimaryLightHighContrast = Color(0xFFA8C8FF)
+internal val surfaceDimLightHighContrast = Color(0xFFC4B5B8)
+internal val surfaceBrightLightHighContrast = Color(0xFFFFF8F8)
+internal val surfaceContainerLowestLightHighContrast = Color(0xFFFFFFFF)
+internal val surfaceContainerLowLightHighContrast = Color(0xFFFDEDEF)
+internal val surfaceContainerLightHighContrast = Color(0xFFEFDFE1)
+internal val surfaceContainerHighLightHighContrast = Color(0xFFE0D1D3)
+internal val surfaceContainerHighestLightHighContrast = Color(0xFFD2C3C5)
+
+internal val primaryDark = Color(0xFFA8C8FF)
+internal val onPrimaryDark = Color(0xFF07305F)
+internal val primaryContainerDark = Color(0xFF254777)
+internal val onPrimaryContainerDark = Color(0xFFD6E3FF)
+internal val secondaryDark = Color(0xFF96D5A6)
+internal val onSecondaryDark = Color(0xFF00391C)
+internal val secondaryContainerDark = Color(0xFF12512E)
+internal val onSecondaryContainerDark = Color(0xFFB2F1C1)
+internal val tertiaryDark = Color(0xFFFFB877)
+internal val onTertiaryDark = Color(0xFF4B2700)
+internal val tertiaryContainerDark = Color(0xFF6B3B03)
+internal val onTertiaryContainerDark = Color(0xFFFFDCC0)
+internal val errorDark = Color(0xFFFFB4AB)
+internal val onErrorDark = Color(0xFF690005)
+internal val errorContainerDark = Color(0xFF93000A)
+internal val onErrorContainerDark = Color(0xFFFFDAD6)
+internal val backgroundDark = Color(0xFF111318)
+internal val onBackgroundDark = Color(0xFFE2E2E9)
+internal val surfaceDark = Color(0xFF191113)
+internal val onSurfaceDark = Color(0xFFEFDFE1)
+internal val surfaceVariantDark = Color(0xFF43474E)
+internal val onSurfaceVariantDark = Color(0xFFC4C6CF)
+internal val outlineDark = Color(0xFF8E9099)
+internal val outlineVariantDark = Color(0xFF43474E)
+internal val scrimDark = Color(0xFF000000)
+internal val inverseSurfaceDark = Color(0xFFEFDFE1)
+internal val inverseOnSurfaceDark = Color(0xFF372E30)
+internal val inversePrimaryDark = Color(0xFF3F5F90)
+internal val surfaceDimDark = Color(0xFF191113)
+internal val surfaceBrightDark = Color(0xFF413739)
+internal val surfaceContainerLowestDark = Color(0xFF140C0E)
+internal val surfaceContainerLowDark = Color(0xFF22191C)
+internal val surfaceContainerDark = Color(0xFF261D20)
+internal val surfaceContainerHighDark = Color(0xFF31282A)
+internal val surfaceContainerHighestDark = Color(0xFF3C3235)
+
+internal val primaryDarkMediumContrast = Color(0xFFCCDDFF)
+internal val onPrimaryDarkMediumContrast = Color(0xFF00254F)
+internal val primaryContainerDarkMediumContrast = Color(0xFF7292C6)
+internal val onPrimaryContainerDarkMediumContrast = Color(0xFF000000)
+internal val secondaryDarkMediumContrast = Color(0xFFACEBBB)
+internal val onSecondaryDarkMediumContrast = Color(0xFF002D14)
+internal val secondaryContainerDarkMediumContrast = Color(0xFF629E73)
+internal val onSecondaryContainerDarkMediumContrast = Color(0xFF000000)
+internal val tertiaryDarkMediumContrast = Color(0xFFFFD4B1)
+internal val onTertiaryDarkMediumContrast = Color(0xFF3C1E00)
+internal val tertiaryContainerDarkMediumContrast = Color(0xFFC28348)
+internal val onTertiaryContainerDarkMediumContrast = Color(0xFF000000)
+internal val errorDarkMediumContrast = Color(0xFFFFD2CC)
+internal val onErrorDarkMediumContrast = Color(0xFF540003)
+internal val errorContainerDarkMediumContrast = Color(0xFFFF5449)
+internal val onErrorContainerDarkMediumContrast = Color(0xFF000000)
+internal val backgroundDarkMediumContrast = Color(0xFF111318)
+internal val onBackgroundDarkMediumContrast = Color(0xFFE2E2E9)
+internal val surfaceDarkMediumContrast = Color(0xFF191113)
+internal val onSurfaceDarkMediumContrast = Color(0xFFFFFFFF)
+internal val surfaceVariantDarkMediumContrast = Color(0xFF43474E)
+internal val onSurfaceVariantDarkMediumContrast = Color(0xFFDADCE5)
+internal val outlineDarkMediumContrast = Color(0xFFAFB2BB)
+internal val outlineVariantDarkMediumContrast = Color(0xFF8D9099)
+internal val scrimDarkMediumContrast = Color(0xFF000000)
+internal val inverseSurfaceDarkMediumContrast = Color(0xFFEFDFE1)
+internal val inverseOnSurfaceDarkMediumContrast = Color(0xFF31282A)
+internal val inversePrimaryDarkMediumContrast = Color(0xFF274878)
+internal val surfaceDimDarkMediumContrast = Color(0xFF191113)
+internal val surfaceBrightDarkMediumContrast = Color(0xFF4C4244)
+internal val surfaceContainerLowestDarkMediumContrast = Color(0xFF0C0607)
+internal val surfaceContainerLowDarkMediumContrast = Color(0xFF241B1E)
+internal val surfaceContainerDarkMediumContrast = Color(0xFF2F2628)
+internal val surfaceContainerHighDarkMediumContrast = Color(0xFF3A3032)
+internal val surfaceContainerHighestDarkMediumContrast = Color(0xFF453B3D)
+
+internal val primaryDarkHighContrast = Color(0xFFEBF0FF)
+internal val onPrimaryDarkHighContrast = Color(0xFF000000)
+internal val primaryContainerDarkHighContrast = Color(0xFFA4C4FB)
+internal val onPrimaryContainerDarkHighContrast = Color(0xFF000B1F)
+internal val secondaryDarkHighContrast = Color(0xFFBFFFCE)
+internal val onSecondaryDarkHighContrast = Color(0xFF000000)
+internal val secondaryContainerDarkHighContrast = Color(0xFF92D1A2)
+internal val onSecondaryContainerDarkHighContrast = Color(0xFF000F04)
+internal val tertiaryDarkHighContrast = Color(0xFFFFEDE0)
+internal val onTertiaryDarkHighContrast = Color(0xFF000000)
+internal val tertiaryContainerDarkHighContrast = Color(0xFFFAB474)
+internal val onTertiaryContainerDarkHighContrast = Color(0xFF160800)
+internal val errorDarkHighContrast = Color(0xFFFFECE9)
+internal val onErrorDarkHighContrast = Color(0xFF000000)
+internal val errorContainerDarkHighContrast = Color(0xFFFFAEA4)
+internal val onErrorContainerDarkHighContrast = Color(0xFF220001)
+internal val backgroundDarkHighContrast = Color(0xFF111318)
+internal val onBackgroundDarkHighContrast = Color(0xFFE2E2E9)
+internal val surfaceDarkHighContrast = Color(0xFF191113)
+internal val onSurfaceDarkHighContrast = Color(0xFFFFFFFF)
+internal val surfaceVariantDarkHighContrast = Color(0xFF43474E)
+internal val onSurfaceVariantDarkHighContrast = Color(0xFFFFFFFF)
+internal val outlineDarkHighContrast = Color(0xFFEEF0F9)
+internal val outlineVariantDarkHighContrast = Color(0xFFC0C2CC)
+internal val scrimDarkHighContrast = Color(0xFF000000)
+internal val inverseSurfaceDarkHighContrast = Color(0xFFEFDFE1)
+internal val inverseOnSurfaceDarkHighContrast = Color(0xFF000000)
+internal val inversePrimaryDarkHighContrast = Color(0xFF274878)
+internal val surfaceDimDarkHighContrast = Color(0xFF191113)
+internal val surfaceBrightDarkHighContrast = Color(0xFF584D50)
+internal val surfaceContainerLowestDarkHighContrast = Color(0xFF000000)
+internal val surfaceContainerLowDarkHighContrast = Color(0xFF261D20)
+internal val surfaceContainerDarkHighContrast = Color(0xFF372E30)
+internal val surfaceContainerHighDarkHighContrast = Color(0xFF43393B)
+internal val surfaceContainerHighestDarkHighContrast = Color(0xFF4F4446)
