+++ composeApp/src/commonTest/kotlin/dev/yuyuyuyuyu/portfolio/ComposeAppCommonTest.kt
@@ -1,0 +1,12 @@
+package dev.yuyuyuyuyu.portfolio
+
+import kotlin.test.Test
+import kotlin.test.assertEquals
+
+class ComposeAppCommonTest {
+
+    @Test
+    fun example() {
+        assertEquals(3, 1 + 2)
+    }
+}
