--- composeApp/src/wasmJsMain/kotlin/dev/yuyuyuyuyu/portfolio/data/repositories/PluginsRepository.kt
@@ -1,23 +1,0 @@
-package dev.yuyuyuyuyu.portfolio.data.repositories
-
-import dev.yuyuyuyuyu.portfolio.data.types.Product
-import dev.yuyuyuyuyu.portfolio.data.types.TechStack
-
-class PluginsRepository {
-    private val plugins = listOf(
-        Product(
-            name = "insert-end-semicolon.nvim",
-            description = "行末に\";\"を挿入するNeovimプラグイン",
-            techStack = setOf(TechStack.Neovim, TechStack.Lua),
-            repositoryUrl = "https://github.com/yu-ko-ba/insert-end-semicolon.nvim",
-        ),
-        Product(
-            name = "ComposePWA",
-            description = "Compose Multiplatform製WebアプリをPWA化するGradleプラグイン",
-            techStack = setOf(TechStack.Kotlin, TechStack.ComposeMultiplatform, TechStack.Gradle),
-            repositoryUrl = "https://github.com/yu-ko-ba/ComposePWA",
-        )
-    )
-
-    fun getPlugins() = plugins
-}
