--- composeApp/src/wasmJsMain/kotlin/dev/yuyuyuyuyu/portfolio/data/types/Product.kt
@@ -1,8 +1,0 @@
-package dev.yuyuyuyuyu.portfolio.data.types
-
-data class Product(
-    val name: String,
-    val description: String,
-    val techStack: Set<TechStack>,
-    val repositoryUrl: String,
-)
