--- composeApp/src/wasmJsMain/kotlin/dev/yuyuyuyuyu/portfolio/data/types/TechStack.kt
@@ -1,33 +1,0 @@
-package dev.yuyuyuyuyu.portfolio.data.types
-
-enum class TechStack(val label: String) {
-    Ios("iOS"),
-    Android("Android"),
-    Web("Web"),
-    Neovim("Neovim"),
-    Gradle("Gradle"),
-    NodeJs("Node.js"),
-    Swift("Swift"),
-    Kotlin("Kotlin"),
-    React("React"),
-    NextJs("Next.js"),
-    Angular("Angular"),
-    TypeScript("TypeScript"),
-    Go("Go"),
-    Lua("Lua"),
-    Sass("Sass"),
-    Html("HTML"),
-    Css("CSS"),
-    JavaScript("JavaScript"),
-    Bash("Bash"),
-    Python("Python"),
-    JetpackCompose("Jetpack Compose"),
-    ComposeMultiplatform("Compose Multiplatform"),
-    MaterialUi("Material UI"),
-    AngularMaterial("Angular Material"),
-    Aws("AWS"),
-    AwsLambda("AWS Lambda"),
-    AmazonDynamoDb("Amazon DynamoDB"),
-    Docker("Docker"),
-    Pwa("Progressive web app (PWA)"),
-}
