--- composeApp/src/wasmJsMain/kotlin/dev/yuyuyuyuyu/portfolio/ui/NavigateDestination.kt
@@ -1,6 +1,0 @@
-package dev.yuyuyuyuyu.portfolio.ui
-
-enum class NavigateDestination {
-    Portfolio,
-    OpenSourceLicenses,
-}
