--- composeApp/src/wasmJsMain/kotlin/dev/yuyuyuyuyu/portfolio/ui/theme/Color.kt
@@ -1,223 +1,0 @@
-package dev.yuyuyuyuyu.portfolio.ui.theme
-
-import androidx.compose.ui.graphics.Color
-
-val primaryLight = Color(0xFF1E6A4F)
-val onPrimaryLight = Color(0xFFFFFFFF)
-val primaryContainerLight = Color(0xFFFFDADC)
-val onPrimaryContainerLight = Color(0xFF3B0712)
-
-val secondaryLight = Color(0xFF765659)
-val onSecondaryLight = Color(0xFFFFFFFF)
-
-val secondaryContainerLight = Color(0xFFD7E3F8)
-val onSecondaryContainerLight = Color(0xFF2C1518)
-
-val tertiaryLight = Color(0xFF785930)
-val onTertiaryLight = Color(0xFFFFFFFF)
-
-val tertiaryContainerLight = Color(0xFFFFDDB6)
-val onTertiaryContainerLight = Color(0xFF2A1800)
-val errorLight = Color(0xFFBA1A1A)
-val onErrorLight = Color(0xFFFFFFFF)
-val errorContainerLight = Color(0xFFFFDAD6)
-val onErrorContainerLight = Color(0xFF410002)
-val backgroundLight = Color(0xFFFFF8F7)
-val onBackgroundLight = Color(0xFF22191A)
-val surfaceLight = Color(0xFFFFF8F7)
-val onSurfaceLight = Color(0xFF22191A)
-val surfaceVariantLight = Color(0xFFF4DDDE)
-val onSurfaceVariantLight = Color(0xFF524344)
-val outlineLight = Color(0xFF847374)
-val outlineVariantLight = Color(0xFFD7C1C2)
-val scrimLight = Color(0xFF000000)
-val inverseSurfaceLight = Color(0xFF382E2F)
-val inverseOnSurfaceLight = Color(0xFFFFEDED)
-val inversePrimaryLight = Color(0xFFFFB2B9)
-val surfaceDimLight = Color(0xFFE7D6D6)
-val surfaceBrightLight = Color(0xFFFFF8F7)
-val surfaceContainerLowestLight = Color(0xFFFFFFFF)
-val surfaceContainerLowLight = Color(0xFFFFF0F0)
-val surfaceContainerLight = Color(0xFFFCEAEA)
-val surfaceContainerHighLight = Color(0xFFF6E4E4)
-val surfaceContainerHighestLight = Color(0xFFF0DEDF)
-
-val primaryLightMediumContrast = Color(0xFF6D2F38)
-val onPrimaryLightMediumContrast = Color(0xFFFFFFFF)
-val primaryContainerLightMediumContrast = Color(0xFFA95F68)
-val onPrimaryContainerLightMediumContrast = Color(0xFFFFFFFF)
-val secondaryLightMediumContrast = Color(0xFF583B3E)
-val onSecondaryLightMediumContrast = Color(0xFFFFFFFF)
-val secondaryContainerLightMediumContrast = Color(0xFF8E6C6F)
-val onSecondaryContainerLightMediumContrast = Color(0xFFFFFFFF)
-val tertiaryLightMediumContrast = Color(0xFF593D17)
-val onTertiaryLightMediumContrast = Color(0xFFFFFFFF)
-val tertiaryContainerLightMediumContrast = Color(0xFF906E44)
-val onTertiaryContainerLightMediumContrast = Color(0xFFFFFFFF)
-val errorLightMediumContrast = Color(0xFF8C0009)
-val onErrorLightMediumContrast = Color(0xFFFFFFFF)
-val errorContainerLightMediumContrast = Color(0xFFDA342E)
-val onErrorContainerLightMediumContrast = Color(0xFFFFFFFF)
-val backgroundLightMediumContrast = Color(0xFFFFF8F7)
-val onBackgroundLightMediumContrast = Color(0xFF22191A)
-val surfaceLightMediumContrast = Color(0xFFFFF8F7)
-val onSurfaceLightMediumContrast = Color(0xFF22191A)
-val surfaceVariantLightMediumContrast = Color(0xFFF4DDDE)
-val onSurfaceVariantLightMediumContrast = Color(0xFF4E3F40)
-val outlineLightMediumContrast = Color(0xFF6B5B5C)
-val outlineVariantLightMediumContrast = Color(0xFF887677)
-val scrimLightMediumContrast = Color(0xFF000000)
-val inverseSurfaceLightMediumContrast = Color(0xFF382E2F)
-val inverseOnSurfaceLightMediumContrast = Color(0xFFFFEDED)
-val inversePrimaryLightMediumContrast = Color(0xFFFFB2B9)
-val surfaceDimLightMediumContrast = Color(0xFFE7D6D6)
-val surfaceBrightLightMediumContrast = Color(0xFFFFF8F7)
-val surfaceContainerLowestLightMediumContrast = Color(0xFFFFFFFF)
-val surfaceContainerLowLightMediumContrast = Color(0xFFFFF0F0)
-val surfaceContainerLightMediumContrast = Color(0xFFFCEAEA)
-val surfaceContainerHighLightMediumContrast = Color(0xFFF6E4E4)
-val surfaceContainerHighestLightMediumContrast = Color(0xFFF0DEDF)
-
-val primaryLightHighContrast = Color(0xFF440E19)
-val onPrimaryLightHighContrast = Color(0xFFFFFFFF)
-val primaryContainerLightHighContrast = Color(0xFF6D2F38)
-val onPrimaryContainerLightHighContrast = Color(0xFFFFFFFF)
-val secondaryLightHighContrast = Color(0xFF331B1E)
-val onSecondaryLightHighContrast = Color(0xFFFFFFFF)
-val secondaryContainerLightHighContrast = Color(0xFF583B3E)
-val onSecondaryContainerLightHighContrast = Color(0xFFFFFFFF)
-val tertiaryLightHighContrast = Color(0xFF331E00)
-val onTertiaryLightHighContrast = Color(0xFFFFFFFF)
-val tertiaryContainerLightHighContrast = Color(0xFF593D17)
-val onTertiaryContainerLightHighContrast = Color(0xFFFFFFFF)
-val errorLightHighContrast = Color(0xFF4E0002)
-val onErrorLightHighContrast = Color(0xFFFFFFFF)
-val errorContainerLightHighContrast = Color(0xFF8C0009)
-val onErrorContainerLightHighContrast = Color(0xFFFFFFFF)
-val backgroundLightHighContrast = Color(0xFFFFF8F7)
-val onBackgroundLightHighContrast = Color(0xFF22191A)
-val surfaceLightHighContrast = Color(0xFFFFF8F7)
-val onSurfaceLightHighContrast = Color(0xFF000000)
-val surfaceVariantLightHighContrast = Color(0xFFF4DDDE)
-val onSurfaceVariantLightHighContrast = Color(0xFF2D2122)
-val outlineLightHighContrast = Color(0xFF4E3F40)
-val outlineVariantLightHighContrast = Color(0xFF4E3F40)
-val scrimLightHighContrast = Color(0xFF000000)
-val inverseSurfaceLightHighContrast = Color(0xFF382E2F)
-val inverseOnSurfaceLightHighContrast = Color(0xFFFFFFFF)
-val inversePrimaryLightHighContrast = Color(0xFFFFE6E7)
-val surfaceDimLightHighContrast = Color(0xFFE7D6D6)
-val surfaceBrightLightHighContrast = Color(0xFFFFF8F7)
-val surfaceContainerLowestLightHighContrast = Color(0xFFFFFFFF)
-val surfaceContainerLowLightHighContrast = Color(0xFFFFF0F0)
-val surfaceContainerLightHighContrast = Color(0xFFFCEAEA)
-val surfaceContainerHighLightHighContrast = Color(0xFFF6E4E4)
-val surfaceContainerHighestLightHighContrast = Color(0xFFF0DEDF)
-
-val primaryDark = Color(0xFFFFB2B9)
-val onPrimaryDark = Color(0xFF561D26)
-val primaryContainerDark = Color(0xFF72333C)
-val onPrimaryContainerDark = Color(0xFFFFDADC)
-val secondaryDark = Color(0xFFE5BDBF)
-val onSecondaryDark = Color(0xFF43292C)
-val secondaryContainerDark = Color(0xFF5C3F42)
-val onSecondaryContainerDark = Color(0xFFFFDADC)
-val tertiaryDark = Color(0xFFE8C08E)
-val onTertiaryDark = Color(0xFF442B06)
-val tertiaryContainerDark = Color(0xFF5D411B)
-val onTertiaryContainerDark = Color(0xFFFFDDB6)
-val errorDark = Color(0xFFFFB4AB)
-val onErrorDark = Color(0xFF690005)
-val errorContainerDark = Color(0xFF93000A)
-val onErrorContainerDark = Color(0xFFFFDAD6)
-val backgroundDark = Color(0xFF1A1112)
-val onBackgroundDark = Color(0xFFF0DEDF)
-val surfaceDark = Color(0xFF1A1112)
-val onSurfaceDark = Color(0xFFF0DEDF)
-val surfaceVariantDark = Color(0xFF524344)
-val onSurfaceVariantDark = Color(0xFFD7C1C2)
-val outlineDark = Color(0xFF9F8C8D)
-val outlineVariantDark = Color(0xFF524344)
-val scrimDark = Color(0xFF000000)
-val inverseSurfaceDark = Color(0xFFF0DEDF)
-val inverseOnSurfaceDark = Color(0xFF382E2F)
-val inversePrimaryDark = Color(0xFF8F4952)
-val surfaceDimDark = Color(0xFF1A1112)
-val surfaceBrightDark = Color(0xFF413737)
-val surfaceContainerLowestDark = Color(0xFF140C0D)
-val surfaceContainerLowDark = Color(0xFF22191A)
-val surfaceContainerDark = Color(0xFF271D1E)
-val surfaceContainerHighDark = Color(0xFF312828)
-val surfaceContainerHighestDark = Color(0xFF3D3233)
-
-val primaryDarkMediumContrast = Color(0xFFFFB8BF)
-val onPrimaryDarkMediumContrast = Color(0xFF34030D)
-val primaryContainerDarkMediumContrast = Color(0xFFCA7A83)
-val onPrimaryContainerDarkMediumContrast = Color(0xFF000000)
-val secondaryDarkMediumContrast = Color(0xFFEAC1C4)
-val onSecondaryDarkMediumContrast = Color(0xFF261012)
-val secondaryContainerDarkMediumContrast = Color(0xFFAC888A)
-val onSecondaryContainerDarkMediumContrast = Color(0xFF000000)
-val tertiaryDarkMediumContrast = Color(0xFFEDC492)
-val onTertiaryDarkMediumContrast = Color(0xFF231300)
-val tertiaryContainerDarkMediumContrast = Color(0xFFAF8A5D)
-val onTertiaryContainerDarkMediumContrast = Color(0xFF000000)
-val errorDarkMediumContrast = Color(0xFFFFBAB1)
-val onErrorDarkMediumContrast = Color(0xFF370001)
-val errorContainerDarkMediumContrast = Color(0xFFFF5449)
-val onErrorContainerDarkMediumContrast = Color(0xFF000000)
-val backgroundDarkMediumContrast = Color(0xFF1A1112)
-val onBackgroundDarkMediumContrast = Color(0xFFF0DEDF)
-val surfaceDarkMediumContrast = Color(0xFF1A1112)
-val onSurfaceDarkMediumContrast = Color(0xFFFFF9F9)
-val surfaceVariantDarkMediumContrast = Color(0xFF524344)
-val onSurfaceVariantDarkMediumContrast = Color(0xFFDBC6C7)
-val outlineDarkMediumContrast = Color(0xFFB29E9F)
-val outlineVariantDarkMediumContrast = Color(0xFF917F80)
-val scrimDarkMediumContrast = Color(0xFF000000)
-val inverseSurfaceDarkMediumContrast = Color(0xFFF0DEDF)
-val inverseOnSurfaceDarkMediumContrast = Color(0xFF312828)
-val inversePrimaryDarkMediumContrast = Color(0xFF74343D)
-val surfaceDimDarkMediumContrast = Color(0xFF1A1112)
-val surfaceBrightDarkMediumContrast = Color(0xFF413737)
-val surfaceContainerLowestDarkMediumContrast = Color(0xFF140C0D)
-val surfaceContainerLowDarkMediumContrast = Color(0xFF22191A)
-val surfaceContainerDarkMediumContrast = Color(0xFF271D1E)
-val surfaceContainerHighDarkMediumContrast = Color(0xFF312828)
-val surfaceContainerHighestDarkMediumContrast = Color(0xFF3D3233)
-
-val primaryDarkHighContrast = Color(0xFFFFF9F9)
-val onPrimaryDarkHighContrast = Color(0xFF000000)
-val primaryContainerDarkHighContrast = Color(0xFFFFB8BF)
-val onPrimaryContainerDarkHighContrast = Color(0xFF000000)
-val secondaryDarkHighContrast = Color(0xFFFFF9F9)
-val onSecondaryDarkHighContrast = Color(0xFF000000)
-val secondaryContainerDarkHighContrast = Color(0xFFEAC1C4)
-val onSecondaryContainerDarkHighContrast = Color(0xFF000000)
-val tertiaryDarkHighContrast = Color(0xFFFFFAF7)
-val onTertiaryDarkHighContrast = Color(0xFF000000)
-val tertiaryContainerDarkHighContrast = Color(0xFFEDC492)
-val onTertiaryContainerDarkHighContrast = Color(0xFF000000)
-val errorDarkHighContrast = Color(0xFFFFF9F9)
-val onErrorDarkHighContrast = Color(0xFF000000)
-val errorContainerDarkHighContrast = Color(0xFFFFBAB1)
-val onErrorContainerDarkHighContrast = Color(0xFF000000)
-val backgroundDarkHighContrast = Color(0xFF1A1112)
-val onBackgroundDarkHighContrast = Color(0xFFF0DEDF)
-val surfaceDarkHighContrast = Color(0xFF1A1112)
-val onSurfaceDarkHighContrast = Color(0xFFFFFFFF)
-val surfaceVariantDarkHighContrast = Color(0xFF524344)
-val onSurfaceVariantDarkHighContrast = Color(0xFFFFF9F9)
-val outlineDarkHighContrast = Color(0xFFDBC6C7)
-val outlineVariantDarkHighContrast = Color(0xFFDBC6C7)
-val scrimDarkHighContrast = Color(0xFF000000)
-val inverseSurfaceDarkHighContrast = Color(0xFFF0DEDF)
-val inverseOnSurfaceDarkHighContrast = Color(0xFF000000)
-val inversePrimaryDarkHighContrast = Color(0xFF4E1620)
-val surfaceDimDarkHighContrast = Color(0xFF1A1112)
-val surfaceBrightDarkHighContrast = Color(0xFF413737)
-val surfaceContainerLowestDarkHighContrast = Color(0xFF140C0D)
-val surfaceContainerLowDarkHighContrast = Color(0xFF22191A)
-val surfaceContainerDarkHighContrast = Color(0xFF271D1E)
-val surfaceContainerHighDarkHighContrast = Color(0xFF312828)
-val surfaceContainerHighestDarkHighContrast = Color(0xFF3D3233)
