+++ settings.gradle.kts
@@ -28,4 +28,8 @@ dependencyResolutionManagement {
     }
 }
 
+plugins {
+    id("org.gradle.toolchains.foojay-resolver-convention") version "0.10.0"
+}
+
 include(":composeApp")
